requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
langchain-groq>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.3.0
langgraph>=0.3.0
duckduckgo-search>=7.0.0
numexpr>=2.9.0
ddgs>=9.13.0